* The application returns the real data * Tests

+++ src/main/java/com/koplisoftl/currency/dto/CustomCurrencyBitcoinPriceIndex.java
@@ -1,22 +1,24 @@
 package com.koplisoftl.currency.dto;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CustomCurrencyBitcoinPriceIndex {
 	@JsonProperty("rate_float")
-	private float rate;
+	private BigDecimal rate;
 	
 	public CustomCurrencyBitcoinPriceIndex() {
 		super();
 	}
 	
-	public CustomCurrencyBitcoinPriceIndex(float rate) {
+	public CustomCurrencyBitcoinPriceIndex(BigDecimal rate) {
 		this.rate = rate;
 	}
 	
-	public float getRate() {
+	public BigDecimal getRate() {
 		return rate;
 	}
 }
